requirements.txt
# Core dependencies
streamlit>=1.28.0
openai>=1.0.0
psycopg2-binary>=2.9.0
pydantic>=2.0.0
requests>=2.31.0

# API layer
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Development
pytest>=7.0.0
python-dotenv>=1.0.0
